Update company-cadence.md (#231)
* Update company-cadence.md

Updated the link to September 2019 and added in March 2020. Also removed the Day from the dates so it's only "Month YYYY" since we don't do it the same day every time. I can plan for that though, if you want to keep days in.

* Update company-cadence.md

Updated March 2019 link to results based on @jasonblais feedback on prior link.

* Correct 2019 with 2020 date

Co-authored-by: Jason Blais <13119842+jasonblais@users.noreply.github.com>