Update recruiting.md
Is it possible to rename this page to "Things Candidates Should Know" ? @jasonblais - I added you as a reviewer as a culture/gut check. Feel free to remove anything that you feel doesn't read properly, including this suggested title. I used it because we have similar pages for Staff and Managers. 

Other changes: 
- Added more info from core handbook (5 mindsets for distributed teams) 
- Changed headers to highlight interview and offer stages 
- Updated FAQ to be General Benefits, a bit more engaging for candidates