Creating an Event Code of Conduct (#785)
* Creating an Event Code of Conduct

This is the addition of the event Code of Conduct to ensure any events we run or are a part of are governed by a set of standards to ensure we are creating a safe, inclusive environments. Attendees of events we organize and Mattermost employees, contributors, and contractors will be held to this standard. This will also be applied to Mattermost employees, contributors, and contractors at any external events.

* Update event-code-of-conduct.md

Minor formatting changes - paragraph breaks.

Co-authored-by: Justine Geffen <justinegeffen@users.noreply.github.com>