Deprecate core committers list
After some reflection on how frequently this document gets out of date, we're proposing simply removing the list altogether from the handbook.

Historically, we have tried many times to keep this document fresh, but as it's not the source of truth, it often gets forgotten. Moving forward, we expect to rely more heavily on Mattermost and GitHub groups to help with discovery issues.