Updated vacation-ready & added video link
A video has been added as a link that will take users with appropriate access to the Mattermost Community Server where the video is hosted.
To embed a video on this page, the video must be publicly available, and can't reside on a server where access is limited. Attempting to embed a video that isn't publicly available will result in no users being able to view the embedded video directly from the Handbook page.