Change "Research and Development" to PDE and consolidate Contributors pages (#1453)
* Change "Research and Development" to PDE

This leaves the 'research-and-development' as part of the URL. Happy to change that out but wanted to avoid a "git wall" for now.

* Remove link to Contributors>Contributors and move children up in TOC

The Contributors > Contributors readme is very out of date and was never really filled out. I propose to just short circut this issue by removing that link from the summary.md page and moving its children up in the Left hand Table of Contents (aka, in the summary.md file). This preserves all the existing URLs.